rename retriever modules
