+++ haystack/retriever/dpr_utils.py
@@ -1,0 +1,97 @@
+#!/usr/bin/env python3
+# Copyright (c) Facebook, Inc. and its affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the license found in the
+# LICENSE file in the root directory of this source tree.
+
+"""
+ Command line tool that produces embeddings for a large documents base based on the pretrained ctx & question encoders
+ Supposed to be used in a 'sharded' way to speed up the process.
+"""
+import argparse
+import csv
+import logging
+import pickle
+from typing import List, Tuple
+
+import numpy as np
+import torch
+from torch import nn
+from torch import Tensor as T
+
+from dpr.models import init_biencoder_components
+from dpr.options import add_encoder_params, setup_args_gpu, print_args, \
+    add_tokenizer_params, add_cuda_params
+from dpr.utils.data_utils import Tensorizer
+from dpr.utils.model_utils import setup_for_distributed_mode, get_model_obj, load_states_from_checkpoint
+
+from dpr.models.hf_models import HFBertEncoder
+
+from transformers.tokenization_bert import BertTokenizer
+
+logger = logging.getLogger(__name__)
+logging.basicConfig(level=logging.INFO)
+
+
+
+
+
+class BertTensorizer(Tensorizer):
+    def __init__(self, tokenizer: BertTokenizer, max_length: int, pad_to_max: bool = True):
+        self.tokenizer = tokenizer
+        self.max_length = max_length
+        self.pad_to_max = pad_to_max
+
+    def text_to_tensor(self, text: str, title: str = None, add_special_tokens: bool = True):
+        text = text.strip()
+
+        # tokenizer automatic padding is explicitly disabled since its inconsistent behavior
+        if title:
+            token_ids = self.tokenizer.encode(title, text_pair=text, add_special_tokens=add_special_tokens,
+                                              max_length=self.max_length,
+                                              pad_to_max_length=False)
+        else:
+            token_ids = self.tokenizer.encode(text, add_special_tokens=add_special_tokens, max_length=self.max_length,
+                                              pad_to_max_length=False)
+
+        seq_len = self.max_length
+        if self.pad_to_max and len(token_ids) < seq_len:
+            token_ids = token_ids + [self.tokenizer.pad_token_id] * (seq_len - len(token_ids))
+        if len(token_ids) > seq_len:
+            token_ids = token_ids[0:seq_len]
+            token_ids[-1] = self.tokenizer.sep_token_id
+
+        return torch.tensor(token_ids)
+
+    def get_pair_separator_ids(self) -> T:
+        return torch.tensor([self.tokenizer.sep_token_id])
+
+    def get_pad_id(self) -> int:
+        return self.tokenizer.pad_token_type_id
+
+    def get_attn_mask(self, tokens_tensor: T) -> T:
+        return tokens_tensor != self.get_pad_id()
+
+    def is_sub_word_id(self, token_id: int):
+        token = self.tokenizer.convert_ids_to_tokens([token_id])[0]
+        return token.startswith("##") or token.startswith(" ##")
+
+    def to_string(self, token_ids, skip_special_tokens=True):
+        return self.tokenizer.decode(token_ids, skip_special_tokens=True)
+
+    def set_pad_to_max(self, do_pad: bool):
+        self.pad_to_max = do_pad
+
+
+
+
+if __name__ == "__main__":
+    r = DPRRetriever()
+    text = """Aaron Aaron ( or ; "Ahärôn") is a prophet, high priest, and the brother of Moses in the Abrahamic religions. Knowledge of Aaron, along with his brother Moses, comes exclusively from religious texts, such as the Bible and Quran. The Hebrew Bible relates that, unlike Moses, who grew up in the Egyptian royal court, Aaron and his elder sister Miriam remained with their kinsmen in the eastern border-land of Egypt (Goshen). When Moses first confronted the Egyptian king about the Israelites, Aaron served as his brother's spokesman ("prophet") to the Pharaoh. Part of the Law (Torah) that Moses received from"""
+    texts = [text]
+    res = r.create_embedding_passage(texts)
+    print(res)
+    print("#################")
+    res = r.create_embedding_query(["Who is Aaron?"])
+    print(res)
