+++ haystack/retriever/sparse.py
@@ -1,11 +1,14 @@
 import logging
-from typing import List, Union
+from collections import OrderedDict
+from typing import List
 
-from farm.infer import Inferencer
+import pandas as pd
+from sklearn.feature_extraction.text import TfidfVectorizer
 
-from haystack.database.base import Document
+from haystack.database.base import Document, BaseDocumentStore
 from haystack.database.elasticsearch import ElasticsearchDocumentStore
 from haystack.retriever.base import BaseRetriever
+from collections import namedtuple
 
 logger = logging.getLogger(__name__)
 
@@ -101,73 +104,95 @@ class ElasticsearchRetriever(BaseRetriever):
         return {"recall": recall, "map": mean_avg_precision}
 
 
-class EmbeddingRetriever(BaseRetriever):
-    def __init__(
-        self,
-        document_store: ElasticsearchDocumentStore,
-        embedding_model: str,
-        gpu: bool = True,
-        model_format: str = "farm",
-        pooling_strategy: str = "reduce_mean",
-        emb_extraction_layer: int = -1,
-    ):
-        """
-        TODO
-        :param document_store:
-        :param embedding_model:
-        :param gpu:
-        :param model_format:
-        """
+# TODO make Paragraph generic for configurable units of text eg, pages, paragraphs, or split by a char_limit
+Paragraph = namedtuple("Paragraph", ["paragraph_id", "document_id", "text", "meta"])
+
+class TfidfRetriever(BaseRetriever):
+    """
+    Read all documents from a SQL backend.
+
+    Split documents into smaller units (eg, paragraphs or pages) to reduce the
+    computations when text is passed on to a Reader for QA.
+
+    It uses sklearn's TfidfVectorizer to compute a tf-idf matrix.
+    """
+
+    def __init__(self, document_store: BaseDocumentStore):
+        self.vectorizer = TfidfVectorizer(
+            lowercase=True,
+            stop_words=None,
+            token_pattern=r"(?u)\b\w\w+\b",
+            ngram_range=(1, 1),
+        )
+
         self.document_store = document_store
-        self.model_format = model_format
-        self.embedding_model = embedding_model
-        self.pooling_strategy = pooling_strategy
-        self.emb_extraction_layer = emb_extraction_layer
+        self.paragraphs = self._get_all_paragraphs()
+        self.df = None
+        self.fit()
 
-        logger.info(f"Init retriever using embeddings of model {embedding_model}")
-        if model_format == "farm" or model_format == "transformers":
-            self.embedding_model = Inferencer.load(
-                embedding_model, task_type="embeddings", extraction_strategy=self.pooling_strategy,
-                extraction_layer=self.emb_extraction_layer, gpu=gpu, batch_size=4, max_seq_len=512, num_processes=0
-            )
+    def _get_all_paragraphs(self) -> List[Paragraph]:
+        """
+        Split the list of documents in paragraphs
+        """
+        documents = self.document_store.get_all_documents()
 
-        elif model_format == "sentence_transformers":
-            from sentence_transformers import SentenceTransformer
+        paragraphs = []
+        p_id = 0
+        for doc in documents:
+            for p in doc.text.split("\n\n"):  # TODO: this assumes paragraphs are separated by "\n\n". Can be switched to paragraph tokenizer.
+                if not p.strip():  # skip empty paragraphs
+                    continue
+                paragraphs.append(
+                    Paragraph(document_id=doc.id, paragraph_id=p_id, text=(p,), meta=doc.meta)
+                )
+                p_id += 1
+        logger.info(f"Found {len(paragraphs)} candidate paragraphs from {len(documents)} docs in DB")
+        return paragraphs
 
-            # pretrained embedding models coming from: https://github.com/UKPLab/sentence-transformers#pretrained-models
-            # e.g. 'roberta-base-nli-stsb-mean-tokens'
-            if gpu:
-                device = "cuda"
-            else:
-                device = "cpu"
-            self.embedding_model = SentenceTransformer(embedding_model, device=device)
-        else:
-            raise NotImplementedError
+    def _calc_scores(self, query: str) -> dict:
+        question_vector = self.vectorizer.transform([query])
 
-    def retrieve(self, query: str, candidate_doc_ids: List[str] = None, top_k: int = 10) -> List[Document]:  # type: ignore
-        query_emb = self.create_embedding(texts=[query])
-        documents = self.document_store.query_by_embedding(query_emb[0], top_k, candidate_doc_ids)
+        scores = self.tfidf_matrix.dot(question_vector.T).toarray()
+        idx_scores = [(idx, score) for idx, score in enumerate(scores)]
+        indices_and_scores = OrderedDict(
+            sorted(idx_scores, key=(lambda tup: tup[1]), reverse=True)
+        )
+        return indices_and_scores
+
+    def retrieve(self, query: str, filters: dict = None, top_k: int = 10, index: str = None) -> List[Document]:
+        if filters:
+            raise NotImplementedError("Filters are not implemented in TfidfRetriever.")
+        if index:
+            raise NotImplementedError("Switching index is not supported in TfidfRetriever.")
+
+        # get scores
+        indices_and_scores = self._calc_scores(query)
+
+        # rank paragraphs
+        df_sliced = self.df.loc[indices_and_scores.keys()]  # type: ignore
+        df_sliced = df_sliced[:top_k]
+
+        logger.debug(
+            f"Identified {df_sliced.shape[0]} candidates via retriever:\n {df_sliced.to_string(col_space=10, index=False)}"
+        )
+
+        # get actual content for the top candidates
+        paragraphs = list(df_sliced.text.values)
+        meta_data = [{"document_id": row["document_id"], "paragraph_id": row["paragraph_id"],  "meta": row.get("meta", {})}
+                     for idx, row in df_sliced.iterrows()]
+
+        documents = []
+        for para, meta in zip(paragraphs, meta_data):
+            documents.append(
+                Document(
+                    id=meta["paragraph_id"],
+                    text=para,
+                    meta=meta.get("meta", {})
+                ))
 
         return documents
 
-    def create_embedding(self, texts: Union[List[str], str]) -> List[List[float]]:
-        """
-        Create embeddings for each text in a list of texts using the retrievers model (`self.embedding_model`)
-        :param texts: texts to embed
-        :return: list of embeddings (one per input text). Each embedding is a list of floats.
-        """
-
-        # for backward compatibility: cast pure str input
-        if type(texts) == str:
-            texts = [texts]  # type: ignore
-        assert type(texts) == list, "Expecting a list of texts, i.e. create_embeddings(texts=['text1',...])"
-
-        if self.model_format == "farm" or self.model_format == "transformers":
-            res = self.embedding_model.inference_from_dicts(dicts=[{"text": t} for t in texts])  # type: ignore
-            emb = [list(r["vec"]) for r in res] #cast from numpy
-        elif self.model_format == "sentence_transformers":
-            # text is single string, sentence-transformers needs a list of strings
-            # get back list of numpy embedding vectors
-            res = self.embedding_model.encode(texts)  # type: ignore
-            emb = [list(r.astype('float64')) for r in res] #cast from numpy
-        return emb
+    def fit(self):
+        self.df = pd.DataFrame.from_dict(self.paragraphs)
+        self.df["text"] = self.df["text"].apply(lambda x: " ".join(x))
+        self.tfidf_matrix = self.vectorizer.fit_transform(self.df["text"])
