+++ test/test_finder.py
@@ -1,7 +1,7 @@
 from haystack import Finder
 from haystack.database.sql import SQLDocumentStore
 from haystack.reader.transformers import TransformersReader
-from haystack.retriever.tfidf import TfidfRetriever
+from haystack.retriever.sparse import TfidfRetriever
 
 
 def test_finder_get_answers():
