+++ tutorials/Tutorial3_Basic_QA_Pipeline_without_Elasticsearch.py
@@ -14,7 +14,7 @@ from haystack.indexing.cleaning import clean_wiki_text
 from haystack.indexing.utils import convert_files_to_dicts, fetch_archive_from_http
 from haystack.reader.farm import FARMReader
 from haystack.reader.transformers import TransformersReader
-from haystack.retriever.tfidf import TfidfRetriever
+from haystack.retriever.sparse import TfidfRetriever
 from haystack.utils import print_answers
