+++ tutorials/Tutorial5_Evaluation.py
@@ -1,6 +1,6 @@
 from haystack.database.elasticsearch import ElasticsearchDocumentStore
 from haystack.indexing.utils import fetch_archive_from_http
-from haystack.retriever.elasticsearch import ElasticsearchRetriever
+from haystack.retriever.sparse import ElasticsearchRetriever
 from haystack.reader.farm import FARMReader
 from haystack.finder import Finder
 from farm.utils import initialize_device_settings
